backend/requirements.txt
Django==3.2.19
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
stripe==5.2.0
sqlparse==0.5.0
PyJWT==2.8.0
cryptography==41.0.5
requests==2.31.0
urllib3==2.0.4
setuptools==68.2.2
wheel==0.41.2
